+++ .dialyzer_ignore.exs
@@ -1,0 +1,5 @@
+# Ignore all Dialyzer warnings for Igniter installer task
+# Mix tasks that use Igniter.Mix.Task have false positives from macros
+[
+  ~r/lib\/mix\/tasks\/beamlens\.install\.ex.*/
+]